requirements.txt
flask==2.3.3
python-dotenv==1.0.0
langchain==0.0.8
faiss-cpu==1.7.2.post2
openai==0.27.0
pickle5==0.0.12
